Rename LineEndingAware* to LineEndingNormalizing* and clean up javadocs
