+++ subprojects/normalization-java/src/main/java/org/gradle/api/internal/changedetection/state/LineEndingNormalizingFileSystemLocationSnapshotHasher.java
@@ -16,13 +16,14 @@
 
 package org.gradle.api.internal.changedetection.state;
 
+import org.gradle.internal.file.FileType;
 import org.gradle.internal.fingerprint.LineEndingSensitivity;
-import org.gradle.internal.fingerprint.hashing.RegularFileSnapshotContext;
+import org.gradle.internal.fingerprint.hashing.FileSystemLocationSnapshotHasher;
 import org.gradle.internal.fingerprint.hashing.ResourceHasher;
-import org.gradle.internal.fingerprint.hashing.ZipEntryContext;
 import org.gradle.internal.hash.HashCode;
 import org.gradle.internal.hash.Hasher;
 import org.gradle.internal.io.IoSupplier;
+import org.gradle.internal.snapshot.FileSystemLocationSnapshot;
 
 import javax.annotation.Nullable;
 import java.io.File;
@@ -30,24 +31,26 @@ import java.io.IOException;
 import java.util.Optional;
 
 /**
- * A {@link ResourceHasher} that ignores line endings while hashing the file.  It detects whether a file is text or binary and only
+ * A {@link FileSystemLocationSnapshotHasher} that normalizes line endings while hashing the file.  It detects whether a file is text or binary and only
  * normalizes line endings for text files.  If a file is detected to be binary, we fall back to the existing non-normalized hash.
+ *
+ * See {@link LineEndingNormalizingInputStreamHasher}.
  */
-public class LineEndingAwareResourceHasher implements ResourceHasher {
-    private final ResourceHasher delegate;
-    private final LineEndingAwareInputStreamHasher hasher;
+public class LineEndingNormalizingFileSystemLocationSnapshotHasher implements FileSystemLocationSnapshotHasher {
+    private final FileSystemLocationSnapshotHasher delegate;
+    private final LineEndingNormalizingInputStreamHasher hasher;
 
-    private LineEndingAwareResourceHasher(ResourceHasher delegate) {
+    private LineEndingNormalizingFileSystemLocationSnapshotHasher(FileSystemLocationSnapshotHasher delegate) {
         this.delegate = delegate;
-        this.hasher = new LineEndingAwareInputStreamHasher();
+        this.hasher = new LineEndingNormalizingInputStreamHasher();
     }
 
-    public static ResourceHasher wrap(ResourceHasher delegate, LineEndingSensitivity lineEndingSensitivity) {
+    public static FileSystemLocationSnapshotHasher wrap(FileSystemLocationSnapshotHasher delegate, LineEndingSensitivity lineEndingSensitivity) {
         switch (lineEndingSensitivity) {
             case DEFAULT:
                 return delegate;
             case NORMALIZE_LINE_ENDINGS:
-                return new LineEndingAwareResourceHasher(delegate);
+                return new LineEndingNormalizingFileSystemLocationSnapshotHasher(delegate);
             default:
                 throw new IllegalArgumentException();
         }
@@ -61,19 +64,12 @@ public class LineEndingAwareResourceHasher implements ResourceHasher {
 
     @Nullable
     @Override
-    public HashCode hash(RegularFileSnapshotContext snapshotContext) throws IOException {
-        return hasher.hashContent(new File(snapshotContext.getSnapshot().getAbsolutePath()))
-            .orElseGet(IoSupplier.wrap(() -> delegate.hash(snapshotContext)));
+    public HashCode hash(FileSystemLocationSnapshot snapshot) throws IOException {
+        return hashContent(snapshot)
+            .orElseGet(IoSupplier.wrap(() -> delegate.hash(snapshot)));
     }
 
-    @Nullable
-    @Override
-    public HashCode hash(ZipEntryContext zipEntryContext) throws IOException {
-        return hashContent(zipEntryContext)
-            .orElseGet(IoSupplier.wrap(() -> delegate.hash(zipEntryContext)));
-    }
-
-    private Optional<HashCode> hashContent(ZipEntryContext zipEntryContext) throws IOException {
-        return zipEntryContext.getEntry().isDirectory() ? Optional.empty() : zipEntryContext.getEntry().withInputStream(hasher::hashContent);
+    private Optional<HashCode> hashContent(FileSystemLocationSnapshot snapshot) throws IOException {
+        return snapshot.getType() == FileType.RegularFile ? hasher.hashContent(new File(snapshot.getAbsolutePath())) : Optional.empty();
     }
 }
